Fixes, moved a file, move code to a file
