+++ frontend/eda/access/credentials/CredentialPage/CredentialTeamAccess.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { TeamAccess } from '../../../common/TeamAccess';
+import { TeamAccess } from '../../common/TeamAccess';
 
 export function CredentialTeamAccess() {
   const params = useParams<{ id: string }>();
