+++ frontend/eda/access/interfaces/TeamAssignment.ts
@@ -1,0 +1,18 @@
+export type TeamAssignment = {
+  id: number;
+  summary_fields: {
+    object_role: {
+      id: number;
+    };
+    role_definition: {
+      id: number;
+      name: string;
+      description: string;
+      managed: boolean;
+    };
+    team: {
+      id: number;
+      name: string;
+    };
+  };
+};
