+++ frontend/eda/decision-environments/DecisionEnvironmentPage/DecisionEnvironmentTeamAccess.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { TeamAccess } from '../../common/TeamAccess';
+import { TeamAccess } from '../../access/common/TeamAccess';
 
 export function DecisionEnvironmentTeamAccess() {
   const params = useParams<{ id: string }>();
