+++ frontend/eda/projects/ProjectPage/ProjectTeamAccess.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { TeamAccess } from '../../common/TeamAccess';
+import { TeamAccess } from '../../access/common/TeamAccess';
 
 export function ProjectTeamAccess() {
   const params = useParams<{ id: string }>();
