+++ frontend/eda/rulebook-activations/RulebookActivationPage/RuleBookActivationTeamAccess.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import { TeamAccess } from '../../common/TeamAccess';
+import { TeamAccess } from '../../access/common/TeamAccess';
 
 export function RulebookActivationTeamAccess() {
   const params = useParams<{ id: string }>();
