+++ frontend/eda/rulebook-activations/RulebookActivationPage/RulebookActivationPage.cy.tsx
@@ -66,11 +66,16 @@ describe('RulebookActivationPage', () => {
   });
 
   it('Should render all the tabs', () => {
-    const tabNames: string[] = ['Back to Rulebook Activations', 'Details', 'History'];
+    const tabNames: string[] = [
+      'Back to Rulebook Activations',
+      'Details',
+      'History',
+      'Team Access',
+    ];
     cy.mount(<RulebookActivationPage />);
 
     cy.get('.pf-v5-c-tabs__list').within(() => {
-      cy.get('.pf-v5-c-tabs__item').should('have.length', 3);
+      cy.get('.pf-v5-c-tabs__item').should('have.length', 4);
       cy.get('.pf-v5-c-tabs__item').each((tab, index) => {
         cy.wrap(tab).should('contain', tabNames[index]);
       });
